discogs_credits: extract Discogs→role mappers
Issue #32 — pure Discogs↔role transforms move to src/mappers.js:

  guessSortName                              "Last, First" / "Name, The"
  getAllArtistTracks                         expand "A1 to A4" ranges
  convertPotentialDJMixers                   route DJ Mix credits per medium
  getArtistRoles                             single-artist role parser
                                             (handles [Co]/[Additional]/...
                                             modifiers, instrument fallback)
  convertDiscogsArtistsToRolesRelationships  flatten over all artists

All pure (or only mutate their direct inputs). Deps are already-
extracted ENTITY_TYPE_MAP, INSTRUMENTS, SELECTORS.

~280 lines moved out. No behavior change; full 12-fixture suite green.
